Generalize logging text/vector segment/index metadata after a flush (#26326)
We're going to care about similar things for text / vector search and I'd like to generalize the remainder of the flushers. So I've created a combined set of metrics that can be used with both.

I ended up expanding the metrics slightly to include both the total number of documents as well as the number of non-deleted documents in each new segment and each index. Total number of documents gives us some idea of how big an index is (actual byte size would be nice too). Non-deleted documents lets us track how much overhead there is for each segment due to deletions.

I also removed logging the size of each segment after a flush. Now we only log the size of the new segment and the total size of the index across all segments. I'll have to fix a few graphs, but this seems more coherent.

GitOrigin-RevId: 1be03cb43651c98477ee7166c829527645da17ce

+++ crates/common/src/knobs.rs
@@ -784,7 +784,7 @@ pub static MULTI_SEGMENT_FULL_SCAN_THRESHOLD_KB: LazyLock<usize> =
 pub static SEGMENT_MAX_SIZE_BYTES: LazyLock<u64> =
     LazyLock::new(|| env_config("SEGMENT_MAX_SIZE_BYTES", (1100000 * 2048 * 4) / 3_u64));
 /// The minimum number of segments we will compact in one pass.
-pub static MIN_COMPACTION_SEGMENTS: LazyLock<usize> =
+pub static MIN_COMPACTION_SEGMENTS: LazyLock<u64> =
     LazyLock::new(|| env_config("MIN_COMPACTION_SEGMENTS", 3));
 /// The maximum percentage of a Segment that can be deleted before we will
 /// recompact that segment to remove deleted vectors
